fix: partial context infos
The previous version of this PR adjusted `withDeclName` to also save the `InfoContext`. This was incorrect because this saved `InfoContext` was not fully complete (e.g. it lacked assignments for metavariables).

The core issue is that we want to save some kind of context while elaborating a term, but cannot meaningfully do so until we are fully done with elaborating the term, lest parts of the saved context may be missing. We already work around this limitation in several other places, e.g. using `Info.updateContext?`, and the parent declaration name is just one more situation where this comes up.

To solve this issue, this commit introduces a notion of `PartialContextInfo` that replaces the `ContextInfo` in `InfoTree.context`. The underlying idea is that while constructing the `InfoTree`, we augment it with `PartialContextInfo`s where it makes sense, and then when traversing over the tree, we merge outer `ContextInfo`s with inner `PartialContextInfo`s (with fields from inner contexts shadowing those of outer context) into `ContextInfo`s to provide all the context that is available up to this point.

The old `ContextInfo` that was added by `liftTermElabM` has been renamed to `CommandContextInfo`.

Another adjustment made in this commit is to remove the incorrect unused variable linter tests checking for unused `where` declarations. `where` declarations are accessible from outside of a definition and should hence not be marked as unused, something that is also fixed by this PR.